refactor(bridge) : div 메서드 return 타입 double -> int 변경

+++ src/main/java/bridge/Calculate.java
@@ -4,5 +4,5 @@ public interface Calculate {
     int add(int a, int b);
     int sub(int a, int b);
     int mul(int a, int b);
-    double div(int a, int b);
+    int div(int a, int b);
 }
